Added frame class to encapsulate frame specific methods. This greatly simplifies the getFrame mess but breaks the API. I'll probably add wrappers to emulate the previous behavior.
git-svn-id: https://ffmpeg-php.svn.sourceforge.net/svnroot/ffmpeg-php/trunk/ffmpeg-php@259 f31480fa-141b-0410-8f85-e1b837e7f62b

+++ tests/getFrame.phpt
@@ -12,16 +12,19 @@ $frame = 73;
 $mov = new ffmpeg_movie(dirname(__FILE__) . '/test_media/test.avi');
 $img = sprintf("%s/test-%04d.png", dirname(__FILE__), $frame);
 
-$gd_image = $mov->getFrame($frame);
-if ($gd_image) {
-    imagepng($gd_image, $img);
-    imagedestroy($gd_image);
-    // generate md5 of file
-    printf("ffmpeg getFrame(): md5 = %s\n", md5(file_get_contents($img)));
-    unlink($img);
-} else {
-    printf("failed\n");
+$ff_frame = $mov->getFrame($frame);
+if ($ff_frame) {
+    $gd_image = $ff_frame->toGDImage();
+    if ($gd_image) {
+        imagepng($gd_image, $img);
+        imagedestroy($gd_image);
+        // generate md5 of file (NOTE: different versions of ffmpeg may produce different
+        // md5 hashes since resampling has been changed slightly due to a fix. Need to
+        // use EXPECTREX to test for both md5 possibilities.
+        printf("ffmpeg getFrame(): md5 = %s\n", md5(file_get_contents($img)));
+        unlink($img);
+    }
 }
 ?>
---EXPECT--
+        --EXPECT--
 ffmpeg getFrame(): md5 = cc36d30fb5b269146802e740ce4ab239
